Making asset_tools/ and CarrotCore compile on Linux

+++ core/core/containers/Vector.ipp
@@ -5,11 +5,12 @@
 #pragma once
 
 #include <algorithm> // for std::sort
+#include <cmath> // for std::ceil
 
 namespace Carrot {
     template<typename TElement>
     constexpr std::size_t getVectorElementAlignment() {
-        return std::max(__STDCPP_DEFAULT_NEW_ALIGNMENT__, alignof(TElement));
+        return std::max(static_cast<std::size_t>(__STDCPP_DEFAULT_NEW_ALIGNMENT__), alignof(TElement));
     }
 
 #define VECTOR_TEMPLATE template<typename TElement, typename Traits> requires IsValidVectorTraits<Traits>
